Fixed bug for announcements that try to happen before DDC starts
Had an error in the code for calling SetTimers(), passing it a full date instead of a difference. Not sure if this is what caused the spamming bug I found earlier, but it's certainly possible. I caught it early and saw it overflowing the timers and setting them to 1 (great idea).